[docs] APISection: move deprecations and platforms to the top of boxes
