+++ docs/components/plugins/api/APISectionConstants.tsx
@@ -4,6 +4,8 @@ import { InlineCode } from '~/components/base/code';
 import { B, P } from '~/components/base/paragraph';
 import { H2, H3Code } from '~/components/plugins/Headings';
 import { ConstantDefinitionData } from '~/components/plugins/api/APIDataTypes';
+import { APISectionDeprecationNote } from '~/components/plugins/api/APISectionDeprecationNote';
+import { PlatformTags } from '~/components/plugins/api/APISectionPlatformTags';
 import {
   CommentTextBlock,
   resolveTypeName,
@@ -20,6 +22,8 @@ const renderConstant = (
   apiName?: string
 ): JSX.Element => (
   <div key={`constant-definition-${name}`} css={STYLES_APIBOX}>
+    <APISectionDeprecationNote comment={comment} />
+    <PlatformTags comment={comment} prefix="Only for:" firstElement />
     <H3Code>
       <InlineCode>
         {apiName ? `${apiName}.` : ''}
@@ -31,7 +35,7 @@ const renderConstant = (
         <B>Type:</B> <InlineCode>{resolveTypeName(type)}</InlineCode>
       </P>
     )}
-    <CommentTextBlock comment={comment} />
+    <CommentTextBlock comment={comment} includePlatforms={false} />
   </div>
 );
